dio sberla

+++ phase2Umps3/ssi.c
@@ -3,12 +3,14 @@
 void klog_print();
 void klog_print_dec();
 void breakPoint();
+extern struct list_head pcbFree_h;
 extern pcb_PTR current_process;
 extern struct list_head *readyQueue;
 extern struct list_head *pseudoClockQueue;
 extern pcb_PTR devQueue[DEVINTNUM][DEVPERINT];
 extern int process_count;
 extern int softBlockCount;
+
 
 void initSSI(){
 	systemServiceInterface();
@@ -25,9 +27,9 @@ void systemServiceInterface(){
 }
 
 void SSIRequest(pcb_t* sender, int service, void* arg){
-	klog_print("SSI request ");
+	/*klog_print("SSI request ");
 	klog_print_dec(service);
-	klog_print("\n");
+	klog_print("\n");*/
 	switch(service){
 		case CREATEPROCESS:
 			createProcess(arg, sender);
@@ -48,7 +50,7 @@ void SSIRequest(pcb_t* sender, int service, void* arg){
 			getSupportStruct(sender);
 			break;
 		case GETPROCESSID:
-			getPID(sender);
+			getPID(arg, sender);
 			break;
 		default:	
 			klog_print("invalid SSI service\n");
@@ -86,14 +88,29 @@ pcb_PTR outDevQ(pcb_PTR doomed){
 
 void killProcess(pcb_PTR doomed, pcb_PTR sender){
 	if(doomed == NULL) doomed = sender;
-	
+
+/*	struct list_head p_list;
+    struct pcb_t *p_parent;
+    struct list_head p_child;
+    struct list_head p_sib;
+
+	p_child list is the sentinel element of p_sib list
+*/
+
+	if(searchProcQ(&pcbFree_h, doomed) == doomed){
+		klog_print("doomed pcb already dead!\n");
+		return;
+	}
+
+	// detach from parent
+	// remove from sibling list
+	outChild(doomed);
+
+	// kill all children
 	while(!emptyChild(doomed)){
 		killProcess(removeChild(doomed), NULL);
 	}
 
-	//remove from sibling list
-	outChild(doomed);
-	
 	//remove from any process queue
 	if(current_process == doomed){
 	}else if(searchProcQ(readyQueue, doomed) == doomed){
@@ -140,9 +157,26 @@ void waitForClock(pcb_PTR sender){
 }
 
 void getSupportStruct(pcb_PTR sender){
-	SYSCALL(SENDMESSAGE, (unsigned int)sender, (unsigned int)sender.p_supportStruct, 0);
+	SYSCALL(SENDMESSAGE, (unsigned int)sender, (unsigned int)sender->p_supportStruct, 0);
 }
 
-void getPID(pcb_PTR sender){
-	SYSCALL(SENDMESSAGE, (unsigned int)sender, (unsigned int)sender->p_pid, 0);
+void getPID(void* arg, pcb_PTR sender){
+/*	unsigned int reply;
+	
+	if( arg==0 ){
+		reply = sender->p_pid;
+	}else if( sender->p_parent==NULL ){
+		reply = 0;
+	}else{
+		reply = sender->p_parent->p_pid;
+	}
+
+	SYSCALL(SENDMESSAGE, (unsigned int)sender, reply, 0);*/
+
+	if( arg==0 )
+		SYSCALL(SENDMESSAGE, (unsigned int)sender, sender->p_pid, 0);
+	else if( sender->p_parent==NULL )
+		SYSCALL(SENDMESSAGE, (unsigned int)sender, 0, 0);
+	else
+		SYSCALL(SENDMESSAGE, (unsigned int)sender, sender->p_parent->p_pid, 0);
 }
